MDEV-15528 Punch holes when pages are freed
When a InnoDB data file page is freed, its contents becomes garbage,
and any storage allocated in the data file is wasted. So InnoDB does
write FREE_PAGE redo record and mark the page status as FREED when
page is being freed or extent is being freed.

If the variable innodb-immediate-scrub-data-uncompressed is true then
write zero out the page. If it is not then ignore the flushing of
the page from flush list.